backend/requirements.txt
google-genai
websockets
pyaudio
python-dotenv
pillow
